Work
Fixed rotation and started working on wires. Looks good so far.

+++ Digital Design CAD Tool/src/components/Line.java
@@ -1,11 +1,15 @@
 package components;
 
 public class Line {
+	// beginning point
 	public double x1;
 	public double y1;
+	
+	// ending point
 	public double x2;
 	public double y2;
-	public Line(int x1, int y1, int x2, int y2) {
+	
+	public Line(double x1, double y1, double x2, double y2) {
 		this.x1 = x1;
 		this.y1 = y1;
 		this.x2 = x2;
